+++ CHANGELOG.md
@@ -2,9 +2,22 @@
 
 ## Unpublished
 
-- Changed cw-multi-test to an in-house cw-multi-test version
+- Added possibility to not panic in parse network by @Buckram123
+- Added stargate feature to cw-multi-test
 - Added `Deploy` to prelude
 - Add ability to provide custom state in `Deploy::set_contracts_state`
+- Breaking change : remove the `&self` dependency for the `Deploy::deployed_state_file_path` method
+- 
+## v0.16.4 [20th September 2023]
+
+- Added automatically derived docs
+- Added gzip support for smaller files sent on-chain
+- Added faster error when broadcast returns a non-0 error code
+
+## v0.16.3 [19th September 2023]
+
+- Changed cw-multi-test to an in-house cw-multi-test version
+- Fixed finding state file using cw-orch-daemon, the Home folder is now recognized from "~"
 
 ## v0.16.2 [12th September 2023]